+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Changelog
   from performance data. (Reported by netphantm and Adrien Nayrat)
   - Support uncompressed files in the archives service. (Suggested by
   Jean-Philippe Guérard)
+  - Add `retention-diff` and `retention-incr` options in the retention service
+  (devopstales).
   - ...